Minimos y maximos por por fecha

La funcion 'min_max_por_estacion()' te devuelve las temperaturas minimas y maximas registradas en cada mes

La funcion 'min_max_por_mes()' te devuelve las temperaturas minimas y maximas registradas en cada mes

Usage

min_max_por_mes(datos)

min_max_por_mes(datos)

Value

Poner la estacion que deseamos ver Ejemplo:

Poner el mes que deseamos ver Ejemplo:

Details

Esta es informacion extra

Esta es informacion extra

Examples

min_max_por_estacion()
#> Error in min_max_por_estacion(): could not find function "min_max_por_estacion"
min_max_por_mes()
#> Error in datos %>% mutate(mes = format(as.Date(fecha), "%Y-%m")) %>% group_by(mes) %>%     summarise(temp_min = min(temperatura, na.rm = TRUE), temp_max = max(temperatura,         na.rm = TRUE)): could not find function "%>%"